+++ release_notes.md
@@ -2,10 +2,13 @@
 
 Under development in `master` branch.
 
+- Renamed main entry point of library from `Processor` to `OverlayJVM`.
+
+
 # 0.1.1
 
-Adding JavaDocs for main classes
+- Adding JavaDocs for main classes
 
 # 0.1.0
 
-First working release
+- First working release